Fix recent access page > Tap on time stamps > App crashes
This bug is a P1 bug blocking our release. It occurs when two preference buttons within the recycler view are clicked simultaneously. The first click successfully navigates to the intended destination, but the second click results in a crash due to a conflict arising from the destination shift caused by the first click.

Bug: 317023097
Test: atest
Change-Id: Ibd94c86985cc85e0318cb15aa78e3ece585e6717