Fix flaky CTS Controller tests.
Recently added test app used exclusively for exercise routes testing (ag/25712909) appears on the screen in other tests (because many tests share the same build target) and pushes some views down outside the visible field of view. Initially it was fixed in ag/25978354 by scrolling to the views but scrolling appeares to be flaky. This change removes the recently added test app and reuses the old test app for testing exercise routes to avoid scrolling.

Bug: 322395395
Test: atest CtsHealthConnectControllerTestCases:ExerciseRouteRequestTest CtsHealthConnectControllerTestCases:PermissionTypesFragmentTest CtsHealthConnectControllerTestCases:ManageAppHealthPermissionUITest CtsHealthConnectControllerTestCases:ManageHealthPermissionsUITest --iterations 5

Change-Id: I4ad9da547491aaae9fd45e77f07390ef1b72b108