Call `initAndValidateUserStateIfNeedLocked` in `FirstGrantTimeManager#setFirstGrantTime`
This would load existing first grant time from disk to the memory cache, then
write the new grant time on top of that. Otherwise, if there are existing
grant times in disk and `setFirstGrantTime` is called before those
grants times are loaded into the memory cache by one of the other
methods such as `getFirstGrantTime` or `onPermissionsChanged`, then
all those grant times on disc will be erased.

By reading the code, the aforementioned case could actually happens with
a delayed migration. `setFirstGrantTime` is used to set grant times
migrated from the APK, so in case of a delayed migration, there could be
existing grant times as the user could have been using HC for a while
before migration happens.

This might or might not be the behavior we want, however, naturally it
would be the right thing to do to load the grant time from disk before
setting a new grant time.

Bug: 321699867
Test: atest "HealthFitnessUnitTests:com.android.server.healthconnect.permission.FirstGrantTimeUnitTest#testSetFirstGrantTimeForAnApp_expectOtherAppsGrantTimesRemained"
Change-Id: Ic37f21f02f3b6ab575510f6f82a6a2c32e47f11b