Fix RejectedExecutionException in FirstGrantTimeManager
The exception could be thrown due to a few reasons:
1. The ThreadPoolExecutor is full, no more tasks can be queued.
2. The ThreadPoolExecutor is shutting down/already terminated.

Looks like crash in the bug was caused by the 2nd reason, so this CL is
catching that exception.

Moreover, the first reason could also potentially happen due to the fact
that onPermissionsChanged could be called rapidly for ALL UIDs on the
device leading to the ThreadPoolExecutor being overloaded, like in this
log https://paste.googleplex.com/5282913497645056. So I moved
the fail fast checks outside of the background task, they don't need to
be run in background anyway.

This CL also adds checks in HealthConnectThreadScheduler to avoid its
methods from throwing RejectedExecutionException if the scheduler is
already shut down.

Test: atest
Fix: 325746130
Change-Id: Ie8e896936ef7f90ab9e7ea6d6bd54bf88df454d1