Add AMRAP and interval based goals.
The AMRAP goal indicates the user should complete as many reps as
possible within the step. This is a performance goal, whereas a fixed
reps goal is a completion goal in its own right. Making it a performance
goal permits flexibility with how the step is completed, e.g. via a
duration, or whenever the user decides to end it (unspecified).

The variable rest goal is typically used in swimming workouts. It has a
specified distance and duration, both of which must be met to complete
the step. Any remaining time after the distance is completed is intended
to be spent resting.

Bug: 319604661
Test: CTS tests
Change-Id: Ie8c788797d8bd9dc6efd3376a6085a8597d75318