Support D2D for training plans data
This introduces merging priority overrides during backup and restore. This enables training plans to be migrated first, followed by exercise sessions, so that a) foreign key constraints can be avoided and b) automatic references between them can be regenerated.

I've also slightly modified upsert transactions. Previously during D2D we were only preserving UUIDs for records where a client ID was specified. This was because UUID generation is deterministic when a client ID is present. For other records, a new, random UUID was generated. Now, the existing UUID is preserved in all cases.

Bug: 319604661
Test: CTS tests
Change-Id: I034b4fb2d32c6bb01b4f72a523c3b33cbb7baab2