Remove Guava usage from system server jar
We're currently inadvertently exposing Guava classes, which can lead to
issues if anything else running in the system server is also referencing
those classes.

The general solution to this is to repackage using
jarjar. However in the case of Guava this leads to a large codesize
increase (140KB), due to keeping of everything annotated with
@VisibleForTesting under com.android. Since our usage of Guava is
trivial, we can instead avoid the problem by just removing it.

Bug: 331282033
Test: atest
Change-Id: I3a1464840ad21542c3969c6fffc98546a72ab886