Bump DB version for training plans release
D2D transfer checks the schema version of the destination instance to
ensure compatible schema versions. Although version 12 and 13 (the
version we bump to in this change) have the
same schema, there is additional business logic related to D2D transfer
that is not present in all versions of HC at schema version 12. This
bump to version 13 ensures that D2D will never attempt to transfer to a
version of HC that doesn't fully support training plans.

Bug: b/319604661
Test: none - configuration change
Change-Id: I64f4bd3376a75c5bf10ae2ca02ae0518a0d5f9df