Sets the "Remove access for all apps" dialog un-cancelable.
The fragment is observing a ViewModel variable for showing or hiding the dialog. We only set the dialog as hidden when either of the buttons have been clicked, not when the user has dismissed the dialog by clicking outside of it.

This change forces users to click on "Cancel" or "Remove all" to dismiss the dialog, and fixes issues where the dialog reappears when re-entering the screen.

Test: manual
Bug: 340497470
Flag: NA
Change-Id: Icae99ea3e0bb58e11162d453893556e637a39848