Query the available document providers.
Used by export/import to show the list of available document providers the user can export/import their data to/from. The query is run on a dedicated executor for the controller app and using a context for the current user so only document providers for the current user are returned.

Bug: 339189778
Test: Integration tests
Flag: com.android.healthfitness.flags.export_import
Change-Id: I81bad274c8dcf19768dbe08dbcd93302c9bd2649