Add an aconfig flag for PHR DB schema changes.
This is a temporary step to unblock PHR implementation while waiting for
the full implementation of go/hc-aconfig-and-db.

With this flag, we can hook PHR APIs to the new DB tables and write
tests for both cases of flag on/off.

When go/hc-aconfig-and-db is implemented, from PHR point of view, it
will just be a refactor, so all PHR test cases should still be
valid by then and no need to be updated.

Test: NA
Bug: 340809376
Flag: com.android.healthfitness.flags.personal_health_record_database
Change-Id: I28ceb67f825ed9d705846a315d766178346f2874