Remove sDatabaseHelpers from DatabaseHelper.
clearAllData in DatabaseHelper relied on the DatabaseHelper constructor
being called to add the instance of a class to the set sDatabaseHelpers.
This caused problems as the data would not be deleted for a particular
class if the instance was not added to the set. Updated the code to
generate the list rather than relying on other classes for calling the
constructor.

Flag: NONE as this is a strict mechanical refactor.
Test: atest HealthFitnessUnitTests
Bug: b/344562759
Change-Id: I9a9e99e5fa9c9d24f8608b882121682fff6c4c9c