RESTRICT AUTOMERGE: Fix change logs related tests
In those affected tests, delete records by IDs instead of time filter.

This is because we have always been generating change logs for delete by
ID API, but we didn't do the same for delete by time filter API until
the bug was fixed.

Now that the bug is fixed, the only way for both old and new code to
pass the tests is to use delete by ID API, where change logs are
generated.

Test: atest CtsHealthFitnessDeviceTestCases
Bug: 345584126
Change-Id: Ia3d8e4188843c085e32b8dc4116f4b40b4529030