"Fix" flaky aggregation test with local time filter
Because of the way HC handles local time in platform, time too close to
EPOCH has negative timestamp representation in some zones.

The prod code doesn't expect any timestamp to be negative, so this was
problematic. This change fixes the test, to unblock the release quickly,
and a follow-up will fix the prod code (with a TODO).

This change also removed debug test case previously added.

Bug: 343425986
Test: atest AggregateWithFiltersTest
Flag: NONE bug fix
Change-Id: I85bb6bb53df4b05aaed829da07aa393b7a2c49ea