Use requireViewById(int) instead of findViewById(int)
Calls to `findViewById(int)` that treat the returned value as if it was
`@NonNull` could result in NPE at runtime as `findViewById(int)` can
return `null`. This change replaces them with calls to the
`requireViewById(int)` method which never returns `null` and will throw
a more useful exception if the view cannot be found.

Flag: EXEMPT refactor
Bug: 347672184
Test: m droid
Change-Id: I57822c195a51cbe075b56ac06eb4d850a9de9f18