Add a test to catch new PHR permissions
If a medical data permission is added, it needs to be added to the category to permission mapping so it is caught by permission enforcement checks. This CL checks all permissions for the "MEDICAL_DATA" substring. Anything with this substring should be returned by the getAllMedicalPermissions method.

Test: atest
Bug: 340204629
Flag: com.android.healthfitness.flags.personal_health_record
Change-Id: Idf26c4b75644b4472a73b1ce7b11658490306412