Add getMedicalPermissionCategory method
In a follow up CL we need to check permissions for the categories of medical data. We don't currently have a good way of mapping from a permission string to the medical category integer. Though we do go the other way around.

This CL creates a new mapping for permission to category so it can be used in the next CL.

Test: atest
Bug: 350435512
Flag: com.android.healthfitness.flags.personal_health_record
Change-Id: I1e7184dff4d3f173582b192f5dd737768bb7c1a2