Refactor: Move static mocking out of HealthConnectDatabaseTestRule and into Tests that use the rule.
This avoids interference when ExtendedMockitoRule and HealthConnectDatabaseTestRule are used in the same test. The problems were likely caused because ExtendedMockitoRule handles Mockito session creation and deletion while HealthConnectDatabaseTestRule was starting a second session concurrently. Required for a test that I am writing, which uses both test rules.

Flag: TEST_ONLY
Test: attest HealthFitnessUnitTests
Bug: 352535790

Change-Id: Iab91326ea02f2aa1883195654a90ebb2aefeed7f