Inline a method throwSecurity().
The method doesn't add anything from clarity, if anything it reduces clarity. I don't think it reduces code size.
If the method is inlined then flow analysis can know that exception throwing definitely happens, which adding the method prevents.
By having this method it forces anything which calls it to be non-static, when that isn't needed.
The best option seems to be remove the method and inline the calls.

Flag: EXEMPT Small code change unlikely to have any impact on droidfood feedback
Test: atest HealthFitnessUnitTests
Change-Id: If1d944b940a4ab7265f9d83dd5c57b288da947fa