Fix PHR unit test failures
The failure was introduced in ag/28679215 which made
`isPersonalHealthRecordEnabled()` take into account
`Flags.developmentDatabase()` as a safety measure to catch the case when
we advance `Flags.personalHealthRecord()` beyond Staging without taking
PHR DB flags into account in `isPersonalHealthRecordEnabled()`.

We already have a TODO in `isPersonalHealthRecordEnabled()` to take PHR
DB flag into account, and it's very unlikely that we'll miss that step.
So removing `Flags.developmentDatabase()` in this CL as a fix.

These failures weren't caught in presubmit because presubmit seems to
only run our tests against trunk_staging builds, so when our unit tests
are not annotated with @EnableFlags(DEVELOPMENT_DATABASE), the default
value of the flag from the device will be used which is `true` for
trunk_staging builds, hence isPersonalHealthRecordEnabled() returns
`true` for the failed unit tests.

Fix: 359400417
Test: atest
Flag: EXEMPT bug fix
Change-Id: Ic6ab9d0412fab4cb2a06b53d498469e1252f8aec