Clear database caches after switching user.
Although the caches are cleared when starting to change user, there's a
race condition with tasks re-populating the caches between clearing the
cache and TransactionManager switching user. This fixes a bug where
PreferenceHelper state is for the wrong user when switching users.

The dependency injection work will improve this so PreferenceHelper
itself will be changed when switching user.

It's not feasible to add a test for this.
HealthConnectManagerServiceTest has lots of mocks and is tricky to
verify the caches are cleared when the race condition is hit. Adding
better tests for handling user switching will be easier once the new
dependency injection work is done.

Bug: 355426144
Test: manual
Flag: com.android.healthfitness.flags.clear_caches_after_switching_user
Change-Id: I506e413f1dc84b7f92127a00ed50f423a1d6cde6