Add tests for export file size logging
For successful export we can get both the exported zip size and by
decompression the db file size. But for an failed export the zip file
does not exist and initial local copies are deleted so have to use
`anyInt()` to match as b/357850072 shows that the file size could differ
on different devices. Using `anyInt()` in this case will make the tests
pass more reliablely.

Test: atest
Flag: EXEMPT tests only
Bug: 357864927
Change-Id: I4e8b781ae7433896bb5255a52270992c22241b11