Add a check to AlterTableRequest to ensure no "NOT NULL" is used
In HC, whenever a module rollback happens, we don't rollback the DB,
hence we are not allowing new columns added to the existing table to be
"NOT NULL". That's because otherwise, when rollback happens, an old
version of the module which is not aware of the new "NOT NULL" columns
won't be able to insert data into that table.

Bug: 340809376
Flag: EXEMPT trivial improvement, no flow change.
Test: unittest
Change-Id: I244d1b9bc2c147a744890b7e7644722d4482e231