Remove some stateful code in DeleteTableRequest that better belongs in TransactionManager.
DeleteTableRequest had a field for number of deleted records and some helper methods. But they were mostly used for an internal loop flow in TransacationManager.
So I've moved them to TransactionManager. Eventually it would be nice for DeleteTableRequest to be immutable, and have a mutable builder, but that is a much bigger change.`

Bug: 353452280
Flag: EXEMPT pure refactor unlikely to cause any visible effect for dogfooders
Test: atest HealthFitnessUnitTests, CtsHealthFitness
Change-Id: I00819120dbfff26832da295d3d1707f4948e196c