Move PreferenceHelper to the Injector. Part One.
PreferenceHelper was being used extensively in
ExportImportSettingsStorage. However, all the functions in
ExportImportSettingsStorage were static to the only way to access the
Injector would be via static reference which I wanted to avoid (as we
want Injector's static access only to be in services). Therefore, I
added PreferenceHelper as a dependency of ExportImportSettingsStorage
and moved ExportImportSettingsStorage as well to the Injector.

It also made sense to move ExportManager to the Injector.

Test: atest HealthFitnessUnitTests
Bug: 362663201
Flag: com.android.healthfitness.flags.dependency_injection
Change-Id: I16a1ba87cd63e424d5652b0640d35ee7d6a341c0