Add new ScheduledExportStatus for error while clearing log tables
The main purpose is being able to identify when export is failing most probably because the file was corrupted during the copy, it being not a valid HealthConnectDatabase anymore. That would help us to make an informed decision about changing or not the current method used for making a local copy of the database.

Bug:b/341890253
Flag: EXEMPT bugfix
Test: atest
Change-Id: I8d1e8896e4640ef15ce0738ead5e88e68fb412c5