Fix test failures due to SecurityException on UDC
On UDC, permissions manager throws a SecurityException if an attempt is
made to grant a permission to an app that hasn't declared it in its
manifest. This behavior was later changed to simply log a warning.

The cause of this crash was a mismatch in declared permissions between
two test apps. We run some code that takes all the granted permissions
of app A and attempt to grant them to app B. Due to the mismatch, this
fails, leading to a SecurityException.

Flag: NONE bugfix
Test: fixing existing test breakage
Bug: b/358261900

Change-Id: If17916fa2ba274b41a5375982ad2847939c23edd