Fix crash loop in TestAppActivity
It appears in at least some cases, where the test app was requested to
kill itself, it was being killed before successfully calling
Activity#finish. It's possible that this lead to the system attempting
to deliver the intent again, thus restarting the app and leading to a
crash loop, which ultimately times out the test.

This attempts to reorder those calls, so that activity#finish is called
up front. Credit to jstembridge@ for the theory and fix proposal.

Bug: 348171256
Test: change to test code itself
Flag: TEST_ONLY
Change-Id: I8a6b12ad785a759a9b420ebff12c275abfb68934